Load and dump YAML as OrderedDict (#3)
By default yaml loads and saves without preserving the order of the dictionary. This makes the result of the vendored snapcraft.yaml less easy to scan for differences. Loader and Dumper need to be overwritten to fix that.

While I was at it, I also used "safe" versions of the respective classes.